Review corrections
-Check pointer before writing to it. It allows application to ask
 only some of the  stagger values.
-Change type of data_amount in get_stagger_estimate_to_address to be
 uint16_t.

+++ features/netsocket/InternetSocket.h
@@ -90,7 +90,9 @@ public:
      *
      * @param address   Destination address to estimate latency.
      * @param latency   Returned latency value in milliseconds.
-     *  @return         NSAPI_ERROR_OK on success, negative error code on failure (@see InternetSocket::getsockopt).
+     *  @return         NSAPI_ERROR_OK on success.
+     *  @return         NSAPI_ERROR_PARAMETER if the provided pointer is invalid.
+     *  @return         negative error code on other failures (@see InternetSocket::getsockopt).
      */
     int get_latency_estimate_to_address(const SocketAddress &address, uint32_t *latency);
 
@@ -101,9 +103,10 @@ public:
      * @param stagger_min   Minimum stagger value in seconds.
      * @param stagger_max   Maximum stagger value in seconds.
      * @param stagger_rand  Randomized stagger value in seconds.
-     * @return              NSAPI_ERROR_OK on success, negative error code on failure (@see InternetSocket::getsockopt).
+     * @return              NSAPI_ERROR_OK on success.
+     * @return              negative error code on other failures (@see InternetSocket::getsockopt).
      */
-    int get_stagger_estimate_to_address(const SocketAddress &address, uint32_t data_amount, uint16_t *stagger_min, uint16_t *stagger_max, uint16_t *stagger_rand);
+    int get_stagger_estimate_to_address(const SocketAddress &address, uint16_t data_amount, uint16_t *stagger_min, uint16_t *stagger_max, uint16_t *stagger_rand);
 
     /** Bind the socket to a port on which to receive data.
      *
